basic restructure done
