+++ src/scenes/landingPage/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+type Props = {}
+
+const LandingPage = (props: Props) => {
+    return (
+        <div>
+            index
+        </div>
+    )
+}
+
+export default LandingPage
