+++ src/scenes/welcomePage/credentialModal/SignUpModal.tsx
