+++ src/scenes/welcomePage/credentialModal/index.tsx
