+++ src/scenes/welcomePage/features/FeatureBox.tsx
